Wrong filenames are fixed
